% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwf_plt_cols.R
\name{fwf_plt_cols}
\alias{fwf_plt_cols}
\title{fwf_plt_cols}
\usage{
fwf_plt_cols(path, cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC"), skip = 6)
}
\arguments{
\item{path}{(character) single path (just one!)}

\item{cols}{(character) select and optionally rename columns of the PLT file}
}
\value{
a tibble, identical in structure to the return value of [readr::fwf_positions()]
}
\description{
Return a column specification suitable for use with [readr::read_fwf()] or [vroom::vroom_fwf()].
}
