% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grf.R
\name{read_grf}
\alias{read_grf}
\title{read_grf}
\usage{
read_grf(path, ..., verbose = getOption("verbose", default = FALSE))
}
\arguments{
\item{path}{(character) ending in .GRF}

\item{...}{passed to [read_plt()]}

\item{verbose}{(logical)}
}
\description{
Treats a .GRF file as a concatenation of .PLT files.
Returns a list where each element is the result of a call to [read_plt()].
Names are extracted from the header chunks in the .GRF file.
}
