% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plt.R
\name{read_plt}
\alias{read_plt}
\title{Import an AERMOD .PLT file as tabular data}
\usage{
read_plt(
  path,
  as = c("tbl", "tbl_cube", "array", "data.table", "raster", "terra", "sf"),
  cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC"),
  complete = TRUE,
  crs = NULL,
  ...,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{path}{(character) path to .PLT file(s)}

\item{as}{(character) format of result}

\item{complete}{(logical) complete with `NA` for (X, Y) combinations that don't appear in the file}

\item{verbose}{(logical)}
}
\value{
an object of type `as` (default: `tbl`)
}
\description{
Import an AERMOD .PLT file as tabular data
}
\note{
`complete = FALSE` only makes sense when `as` is "tbl" or "data.table".
}
\seealso{
predecessor `read_AERMOD_plt.R` in West Oakland AB617 work
}
