% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plt.R
\name{read_plt_as_raster}
\alias{read_plt_as_raster}
\title{read_plt_as_raster}
\usage{
read_plt_as_raster(
  path,
  extent,
  crs = NULL,
  cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC"),
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{path}{(character) filesystem path(s) (see Note)}

\item{extent}{a [raster::Extent]}

\item{crs}{a coordinate reference system (as yielded by [sp::CRS()])}

\item{cols}{(character) as in [read_plt()]}

\item{verbose}{(logical)}
}
\value{
A [raster::RasterLayer] or [raster::RasterStack] (see Note)
}
\description{
read_plt_as_raster
}
\note{
If `path` has more than one element, then the return value will be a [raster::RasterStack], with layer names
    set equal to the [basename()] of each path. Otherwise, the return value will be a [raster::RasterLayer].
}
