% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2cube.R
\name{xy2cube.data.table}
\alias{xy2cube.data.table}
\title{xy2cube.data.table}
\usage{
xy2cube.data.table(xy_data, value_var = "CONC", careful = FALSE)
}
\arguments{
\item{xy_data}{a `data.table` object with columns `X`, `Y`, and whatever `value_var` is}

\item{value_var}{(character) name of column containing values}

\item{careful}{(logical) perform additional QA on-the-fly}
}
\description{
xy2cube.data.table
}
